+++ package.json
@@ -14,7 +14,8 @@
     "@ethereum-attestation-service/eas-sdk": "^1.1.0-beta.2",
     "axios": "^1.4.0",
     "ethers": "5.7",
-    "simple-ts-job-runner": "^1.0.12"
+    "simple-ts-job-runner": "^1.0.12",
+    "viem": "^1.6.4"
   },
   "devDependencies": {
     "@types/node": "^20.5.0",
